+++ src/main/java/com/git_check/member/auth/adapter/in/config/OAuth2LoginComponentConfig.java
@@ -1,4 +1,4 @@
-package com.git_check.member.auth.adapter.config;
+package com.git_check.member.auth.adapter.in.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Bean;
